feat: Introduce ProjectContext for state management
Refactor state management to utilize a new `ProjectContext`. This change centralizes project data, including markdown content, user lists, settings, and project configurations, within a provider.

Components like `UserManagement`, `ProjectOverview`, and `ProjectActions` are updated to consume this context, removing prop drilling and simplifying state access. This promotes a more scalable and maintainable architecture for handling application-wide state.
